% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us-map.R
\name{centroid_labels}
\alias{centroid_labels}
\title{Retrieve centroid labels}
\usage{
centroid_labels(regions = c("states", "counties"), as_sf = TRUE)
}
\arguments{
\item{regions}{The region breakdown for the map, can be one of
(\code{"states"}, \code{"counties"}, as specified by the internal file names.
The default is \code{"states"}.}

\item{as_sf}{Defunct, this parameter no longer has any effect and will be removed in
the future.}
}
\value{
An \code{sf} data frame of state or county centroid labels and positions
relative to the coordinates returned by the \code{us_map} function.
}
\description{
Retrieve centroid labels
}
