% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county}
\alias{county}
\title{US county map data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97788 rows and 7 columns.}
\source{
\url{http://www.census.gov/geo/maps-data/data/cbf/cbf_counties.html}
}
\usage{
county
}
\description{
US county map data form Census.
}
\details{
A data frame with seven variables.
}
\keyword{datasets}

