% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-map-df.R
\name{create_us_map}
\alias{create_us_map}
\title{Create \code{usmap} shapefiles}
\usage{
create_us_map(type = c("states", "counties"), input_file, output_file)
}
\description{
Internal function to create the modified shapefiles used by the
\link[usmap]{usmap} package. Using this function externally is not
recommended since it makes certain undocumented assumptions that may
not work with all inputs.
}
\details{
It is strongly recommend that the \link[usmap]{usmap} package is used
directly.

In some cases where the raw data is required, the \link{us_map}
function located in this package can be used instead.
}
\references{
{
Rudis B (2014). “Moving The Earth
(well, Alaska & Hawaii) With R.”
\url{https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/}.
}
}
\keyword{internal}
