% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{.pkg_env}
\alias{.pkg_env}
\title{Internal package environment for usmapdata.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.pkg_env
}
\description{
Internal package environment for usmapdata.
}
\keyword{internal}
