% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us-map.R
\name{select_map_year}
\alias{select_map_year}
\title{Select appropriate map data year from available years}
\usage{
select_map_year(data_year)
}
\arguments{
\item{data_year}{The year for which to obtain \code{usmap} data.
If the value is \code{NULL}, the most recent year is returned. If the
provided year is not found from the available map data sets, the next most
recent available year is returned.}
}
\description{
Select appropriate map data year from available years
}
\keyword{internal}
