% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-us-map.R
\name{transform}
\alias{transform}
\title{2D transformation matrix}
\usage{
transform(angle = 0, scale = 1)
}
\arguments{
\item{angle}{The rotation angle of the transformation.}

\item{scale}{The scale factor of the transformation.}
}
\value{
A two dimensional transformation matrix for the given
rotation angle and scale factor.
}
\description{
2D transformation matrix
}
\keyword{internal}
