% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips-data.R
\name{fips_data}
\alias{fips_data}
\title{Retrieve state and county FIPS codes}
\usage{
fips_data(regions = c("states", "state", "counties", "county"), as_sf = TRUE)
}
\arguments{
\item{regions}{The region breakdown for the map, can be one of
(\code{"states"}, \code{"state"}, \code{"counties"}, \code{"county"}).
The default is \code{"states"}.}

\item{as_sf}{Defunct, this parameter no longer has any effect and will be removed in
the future.}
}
\value{
An data frame of FIPS codes of the desired \code{regions}.
}
\description{
Retrieve state and county FIPS codes
}
\examples{
str(fips_data())

state_fips <- fips_data()
county_fips <- fips_data(regions = "counties")

}
